+++ doc/SonarLint/SonarLint_complexitas_csokkentes_H9PB1F.md
@@ -1,0 +1,3 @@
+Hérincs Bence Attila
+
+Kaptam egy olyan issue-t, melyben az alébbi 3 helyen csökkentem kell a kognitív komplexitást: OtherAttackActionListner, OtherStealActionListener, ApplyEffect. Erről a feladatkörről már olvastam egy előző issue-nál, így ebben az md fájlban részletesen nem fejteném ki, csak amennyiben valami újat csinálok.
